refactor: separate public and sales views
