+++ application_form/urls.py
@@ -1,11 +1,11 @@
 from django.urls import include, path
 from rest_framework.routers import DefaultRouter
 
-from application_form.api.views import (
-    ApplicationViewSet,
+from application_form.api.sales.views import (
     execute_lottery_for_project,
     SalesApplicationViewSet,
 )
+from application_form.api.views import ApplicationViewSet
 
 router = DefaultRouter()
 router.register(r"applications", ApplicationViewSet)
